pyproject.toml
[project]
name = "quizy-shared"
version = "0.1.0"
description = "Shared lib for quizy services"
readme = "README.md"
authors = [
    { name = "ltmaksb", email = "mbychyniuk@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
